requirements.txt
# Discord Publish Bot - Dependencies
# Managed with UV for fast, reliable package management
# Install with: uv pip install -r requirements.txt

# Discord Bot Dependencies
discord.py==2.3.2
aiohttp==3.9.1

# Publishing API Dependencies  
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==1.59.1
python-multipart==0.0.6
pydantic==2.5.0

# Shared Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0

# Security and Monitoring
slowapi==0.1.9
prometheus-client==0.19.0

# Optional: Enhanced error tracking
# sentry-sdk[fastapi]==1.38.0
